Latest Analysis of AHOSc before investigating getting slower currents
This now includes analysis of the fast reacting (ThinFET based) AH
Oscillator.  It also includes RMSE measurements for the accuracy of the
linear fit of the data points.  Lastly it includes analysis over
temperature and current (0C-50C, 10pA to 10nA).

After this commit, I will start to investigate how the circuit behaves
with slower currents (10fA to 10pA).  We are hoping to see that using a
ThickFET inverter whose input is V_cap, will allow us to do this without
suffering from temperature based performance.