Axon Hillock characterization notebook with latest analysis
This commit tracks the state of the AHOsc Characterization notebook as
of 10.23.15, where we have tested many different configurations of the
AH circuit and are trying to currently solve the temperature dependence
of the circuit at very small currents (100fA).  As of this commit, we
have analyzed the circuit and know that we are getting 6% variation in
output period, over the full 50C span of temperature for 100fA of input
current.

The next step after this commit is to further test whether this error is
coming from the dependence of early voltage upon temperature, caused by
gain in the current mirrors.