AHOsc_Char as of latest working design (using Vcapbiased circuit)
This commit includes the state of my Axon Hillock characterization
including the latest working circuit that uses the new Vcapbiased
configuration to mirror current into the circuit.  The next test after
this commit will be to test out Ben's Mirror to try and see if that
helps my performance while trying to improve the speed of the circuit.